delete goal refactored according to comments

+++ src/main/java/school/faang/user_service/controller/goal/GoalController.java
@@ -1,19 +1,17 @@
 package school.faang.user_service.controller.goal;
 
-import jakarta.validation.constraints.NotEmpty;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import school.faang.user_service.dto.goal.GoalDto;
-import school.faang.user_service.dto.goal.GoalFilterDto;
+import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.service.goal.GoalService;
-
-import java.util.List;
+import school.faang.user_service.util.Message;
 
 @RestController
 @RequiredArgsConstructor
@@ -26,8 +24,6 @@ public class GoalController {
 
         return goalService.createGoal(goalDto, userId);
     }
-}
-
 
     @PutMapping
     public GoalDto updateGoal(@RequestBody GoalDto goalDto, Long userId){
@@ -35,9 +31,13 @@ public class GoalController {
         return goalService.updateGoal(goalDto, userId);
     }
 
-    @DeleteMapping
-    public GoalDto deleteGoal(Long goalId){
-        return goalService.deleteGoal(goalId);
+    @DeleteMapping("/{goalId}")
+    public void deleteGoal(@PathVariable Long goalId){
+        if (goalId < 0){
+            throw new DataValidationException(Message.GOAL_NOT_FOUND);
+        }
+
+        goalService.deleteGoal(goalId);
     }
 
     @GetMapping
